xml_viewsets/requirements.txt
Django==2.0
beautifulsoup4==4.6.3
certifi==2018.8.24
chardet==3.0.4
djangorestframework==3.8.2
idna==2.7
lxml==4.2.5
pytz==2018.5
requests==2.19.1
urllib3==1.23
utils==0.9.0
